CapsGNN/requirements.txt
networkx==1.11
tqdm==4.28.1
numpy==1.15.4
pandas==0.23.4
texttable==1.5.0
scipy==1.1.0
argparse==1.1.0
torch-scatter==1.1.2
torch-sparse==0.2.2
torch-cluster==1.2.4
torch-geometric==1.0.3
torchvision==0.2.1
